3.To

+++ 0x10-variadic_functions/3-print_all.c
@@ -36,7 +36,7 @@ void print_all(const char * const format, ...)
 				printf("%d", va_arg(valist, int));
 				break;
 			case 'f':
-				printf("%f", va_arg(valist, float));
+				printf("%f", va_arg(valist, double));
 				break;
 			case 's':
 				str = va_arg(valist, char *);
